Updated to Swift 4.2 package manager
Also updated Interval : Hashable to latest conformance, which uses `Hasher` to combine hash values.  This means that we can no longer expect a certain value from interval.hashValue and thus the test "has a hashValue" needed to modified slightly.

Updated the CocoaPods version to 1.0.1 due to the slight change to Interval's Hashable conformance.